Revert "GOVSI-748: create separate frontend api"
