+++ ci/terraform/oidc/api-gateway.tf
@@ -121,24 +121,40 @@ resource "aws_api_gateway_deployment" "deployment" {
       module.auth-code.method_trigger_value,
       module.authorize.integration_trigger_value,
       module.authorize.method_trigger_value,
+      module.client-info.integration_trigger_value,
+      module.client-info.method_trigger_value,
       module.jwks.integration_trigger_value,
       module.jwks.method_trigger_value,
+      module.login.integration_trigger_value,
+      module.login.method_trigger_value,
       module.logout.integration_trigger_value,
       module.logout.method_trigger_value,
+      module.mfa.integration_trigger_value,
+      module.mfa.method_trigger_value,
       module.openid_configuration_discovery.integration_trigger_value,
       module.openid_configuration_discovery.method_trigger_value,
       module.register.integration_trigger_value,
       module.register.method_trigger_value,
       module.reset_password.integration_trigger_value,
       module.reset_password.method_trigger_value,
+      module.send_notification.integration_trigger_value,
+      module.send_notification.method_trigger_value,
+      module.signup.integration_trigger_value,
+      module.signup.method_trigger_value,
       module.token.integration_trigger_value,
       module.token.method_trigger_value,
-      module.trustmarks.integration_trigger_value,
-      module.trustmarks.method_trigger_value,
       module.update.integration_trigger_value,
       module.update.method_trigger_value,
+      module.update_profile.integration_trigger_value,
+      module.update_profile.method_trigger_value,
+      module.userexists.integration_trigger_value,
+      module.userexists.method_trigger_value,
       module.userinfo.integration_trigger_value,
       module.userinfo.method_trigger_value,
+      module.verify_code.integration_trigger_value,
+      module.verify_code.method_trigger_value,
+      module.reset-password-request.integration_trigger_value,
+      module.reset-password-request.method_trigger_value
     ]))
   }
 
@@ -148,14 +164,23 @@ resource "aws_api_gateway_deployment" "deployment" {
   depends_on = [
     module.auth-code,
     module.authorize,
+    module.client-info,
     module.jwks,
+    module.login,
     module.logout,
+    module.mfa,
     module.openid_configuration_discovery,
     module.register,
+    module.reset_password,
+    module.send_notification,
+    module.signup,
     module.token,
-    module.trustmarks,
     module.update,
+    module.update_profile,
+    module.userexists,
     module.userinfo,
+    module.verify_code,
+    module.reset-password-request
   ]
 }
 
@@ -169,14 +194,22 @@ resource "aws_api_gateway_stage" "endpoint_stage" {
   depends_on = [
     module.auth-code,
     module.authorize,
+    module.client-info,
     module.jwks,
+    module.login,
     module.logout,
+    module.mfa,
     module.openid_configuration_discovery,
     module.register,
+    module.reset_password,
+    module.send_notification,
+    module.signup,
     module.token,
-    module.trustmarks,
     module.update,
+    module.update_profile,
+    module.userexists,
     module.userinfo,
+    module.verify_code,
     aws_api_gateway_deployment.deployment,
   ]
 }
