+++ ci/terraform/oidc/client-info.tf
@@ -7,7 +7,7 @@ module "client-info" {
   environment     = var.environment
 
   handler_environment_variables = {
-    BASE_URL             = local.frontend_api_base_url
+    BASE_URL             = local.api_base_url
     EVENTS_SNS_TOPIC_ARN = aws_sns_topic.events.arn
     LOCALSTACK_ENDPOINT  = var.use_localstack ? var.localstack_endpoint : null
     REDIS_HOST           = local.external_redis_host
@@ -17,13 +17,13 @@ module "client-info" {
     ENVIRONMENT          = var.environment
     DYNAMO_ENDPOINT      = var.use_localstack ? var.lambda_dynamo_endpoint : null
   }
-  handler_function_name = "uk.gov.di.authentication.frontendapi.lambda.ClientInfoHandler::handleRequest"
+  handler_function_name = "uk.gov.di.authentication.clientregistry.lambda.ClientInfoHandler::handleRequest"
 
-  rest_api_id               = aws_api_gateway_rest_api.di_authentication_frontend_api.id
-  root_resource_id          = aws_api_gateway_rest_api.di_authentication_frontend_api.root_resource_id
-  execution_arn             = aws_api_gateway_rest_api.di_authentication_frontend_api.execution_arn
+  rest_api_id               = aws_api_gateway_rest_api.di_authentication_api.id
+  root_resource_id          = aws_api_gateway_rest_api.di_authentication_api.root_resource_id
+  execution_arn             = aws_api_gateway_rest_api.di_authentication_api.execution_arn
   api_deployment_stage_name = var.api_deployment_stage_name
-  lambda_zip_file           = var.frontend_api_lambda_zip_file
+  lambda_zip_file           = var.client_registry_api_lambda_zip_file
   security_group_id         = local.authentication_security_group_id
   subnet_id                 = local.authentication_subnet_ids
   lambda_role_arn           = local.lambda_iam_role_arn
@@ -39,7 +39,7 @@ module "client-info" {
   use_localstack = var.use_localstack
 
   depends_on = [
-    aws_api_gateway_rest_api.di_authentication_frontend_api,
+    aws_api_gateway_rest_api.di_authentication_api,
     aws_api_gateway_resource.connect_resource,
     aws_api_gateway_resource.wellknown_resource,
   ]
