+++ ci/terraform/oidc/reset-password-request.tf
@@ -8,7 +8,7 @@ module "reset-password-request" {
 
   handler_environment_variables = {
     ENVIRONMENT          = var.environment
-    BASE_URL             = local.frontend_api_base_url
+    BASE_URL             = local.api_base_url
     SQS_ENDPOINT         = var.use_localstack ? "http://localhost:45678/" : null
     EMAIL_QUEUE_URL      = aws_sqs_queue.email_queue.id
     EVENTS_SNS_TOPIC_ARN = aws_sns_topic.events.arn
@@ -21,9 +21,9 @@ module "reset-password-request" {
   }
   handler_function_name = "uk.gov.di.authentication.frontendapi.lambda.ResetPasswordRequestHandler::handleRequest"
 
-  rest_api_id               = aws_api_gateway_rest_api.di_authentication_frontend_api.id
-  root_resource_id          = aws_api_gateway_rest_api.di_authentication_frontend_api.root_resource_id
-  execution_arn             = aws_api_gateway_rest_api.di_authentication_frontend_api.execution_arn
+  rest_api_id               = aws_api_gateway_rest_api.di_authentication_api.id
+  root_resource_id          = aws_api_gateway_rest_api.di_authentication_api.root_resource_id
+  execution_arn             = aws_api_gateway_rest_api.di_authentication_api.execution_arn
   api_deployment_stage_name = var.api_deployment_stage_name
   lambda_zip_file           = var.frontend_api_lambda_zip_file
   security_group_id         = local.authentication_security_group_id
@@ -41,7 +41,7 @@ module "reset-password-request" {
   use_localstack = var.use_localstack
 
   depends_on = [
-    aws_api_gateway_rest_api.di_authentication_frontend_api,
+    aws_api_gateway_rest_api.di_authentication_api,
     aws_api_gateway_resource.connect_resource,
     aws_api_gateway_resource.wellknown_resource,
     aws_sqs_queue.email_queue,
