+++ integration-tests/src/test/java/uk/gov/di/authentication/api/IntegrationTestEndpoints.java
@@ -13,19 +13,9 @@ public class IntegrationTestEndpoints {
             Optional.ofNullable(System.getenv().get("API_GATEWAY_ID")).orElse("");
     protected static final String API_KEY =
             Optional.ofNullable(System.getenv().get("API_KEY")).orElse("");
-    protected static final String FRONTEND_API_KEY =
-            Optional.ofNullable(System.getenv().get("FRONTEND_API_KEY")).orElse("");
     public static final String ROOT_RESOURCE_URL =
             Optional.ofNullable(System.getenv().get("ROOT_RESOURCE_URL"))
                     .orElse(String.format(LOCAL_ENDPOINT_FORMAT, LOCAL_API_GATEWAY_ID));
-    public static final String FRONTEND_ROOT_RESOURCE_URL =
-            Optional.ofNullable(System.getenv().get("ROOT_RESOURCE_URL"))
-                    .orElse(
-                            String.format(
-                                    LOCAL_ENDPOINT_FORMAT,
-                                    Optional.ofNullable(
-                                                    System.getenv().get("FRONTEND_API_GATEWAY_ID"))
-                                            .orElse("")));
 
     @BeforeEach
     public void flushData() {
