+++ integration-tests/src/test/java/uk/gov/di/authentication/api/ResetPasswordRequestIntegrationTest.java
@@ -54,13 +54,10 @@ public class ResetPasswordRequestIntegrationTest extends IntegrationTestEndpoint
         RedisHelper.setSessionState(sessionId, AUTHENTICATION_REQUIRED);
         MultivaluedMap<String, Object> headers = new MultivaluedHashMap<>();
         headers.add("Session-Id", sessionId);
-        headers.add("X-API-Key", FRONTEND_API_KEY);
+        headers.add("X-API-Key", API_KEY);
         Response response =
                 RequestHelper.request(
-                        FRONTEND_ROOT_RESOURCE_URL,
-                        RESET_PASSWORD_ENDPOINT,
-                        new ResetPasswordRequest(email),
-                        headers);
+                        RESET_PASSWORD_ENDPOINT, new ResetPasswordRequest(email), headers);
 
         notifyStub.waitForRequest(60);
 
@@ -84,14 +81,11 @@ public class ResetPasswordRequestIntegrationTest extends IntegrationTestEndpoint
         RedisHelper.setSessionState(sessionId, NEW);
         MultivaluedMap<String, Object> headers = new MultivaluedHashMap<>();
         headers.add("Session-Id", sessionId);
-        headers.add("X-API-Key", FRONTEND_API_KEY);
+        headers.add("X-API-Key", API_KEY);
 
         Response response =
                 RequestHelper.request(
-                        FRONTEND_ROOT_RESOURCE_URL,
-                        RESET_PASSWORD_ENDPOINT,
-                        new ResetPasswordRequest(email),
-                        headers);
+                        RESET_PASSWORD_ENDPOINT, new ResetPasswordRequest(email), headers);
 
         assertEquals(400, response.getStatus());
     }
