+++ integration-tests/src/test/java/uk/gov/di/authentication/helpers/RequestHelper.java
@@ -3,6 +3,7 @@ package uk.gov.di.authentication.helpers;
 import jakarta.ws.rs.client.ClientBuilder;
 import jakarta.ws.rs.client.Entity;
 import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.MultivaluedHashMap;
 import jakarta.ws.rs.core.MultivaluedMap;
 import jakarta.ws.rs.core.Response;
 
@@ -10,18 +11,16 @@ import static uk.gov.di.authentication.api.IntegrationTestEndpoints.ROOT_RESOURC
 
 public class RequestHelper {
 
-    public static Response request(
-            String endpoint, Object body, MultivaluedMap<String, Object> headers) {
-        return request(ROOT_RESOURCE_URL, endpoint, body, headers);
+    public static Response requestWithSession(String endpoint, Object body, String sessionId) {
+        MultivaluedMap<String, Object> headers = new MultivaluedHashMap<>();
+        headers.add("Session-Id", sessionId);
+        return request(endpoint, body, headers);
     }
 
     public static Response request(
-            String rootResourceURL,
-            String endpoint,
-            Object body,
-            MultivaluedMap<String, Object> headers) {
+            String endpoint, Object body, MultivaluedMap<String, Object> headers) {
         return ClientBuilder.newClient()
-                .target(rootResourceURL + endpoint)
+                .target(ROOT_RESOURCE_URL + endpoint)
                 .request(MediaType.APPLICATION_JSON)
                 .headers(headers)
                 .post(Entity.entity(body, MediaType.APPLICATION_JSON));
